requirements.txt
matplotlib==3.3.2
numpy==1.19.2
pandas==1.3.4
plotly==4.13.0
python_dateutil==2.8.2
questionary==1.5.2
scikit_learn==1.0.2
selenium==3.141.0
setuptools==50.3.1
SQLAlchemy==1.3.20
streamlit==1.7.0
tensorflow==2.7.0
webdriver-manager==3.5.4
